Changed Faceted to be an Eq instance
We made Faceted an instance of the Eq typeclass so that we can compare
faceted values for equality. However, this meant we had to constrain the
type Faceted takes in to also be an instance of the Eq typeclass.
Because monads are requried to be fully polymorphic, we used the
workaround shown here
(https://hackage.haskell.org/package/set-monad-0.2.0.0/src/Data/Set/Monad.hs)
so that we can still use monads for faceted values.

I haven't finished implementing the change because of a complication
with FIO. In the function `prod` in FIO.hs, we take in an argument of
type `Faceted (FIO (Faceted a))`. This would require FIO to be an Eq
instance as well so that it can be put inside a faceted value. However,
FIO is defined as `data FIO a = FIO { runFIO :: PC -> IO a }` in
Internal.hs. We haven't yet figured out a way to check the equality of 2
different runFIO functions.