Comment out FHandle and make tests typecheck
We couldn't make `prod` in FIO.hs typecheck since it takes in an
argument of type `Faceted (FIO (Faceted a))`. We can't check the
equality of 2 different runFIO functions, so I just commented out `prod`
and functions that use it in FIO.hs. We also no longer have FHandle,
since it needs `prod` to be able to transform a Faceted (IO a) to IO
(Faceted a). The tests now typecheck, but we still need a way to figure
out how to do faceted IO computations.